Migrate/ts (#123)
* Migrated to typescript
   - Also, generate `flow`.
* Update pod
* Release 2.4.0
