+++ .gitignore
@@ -50,3 +50,9 @@ buck-out/
 Example/android/java_*
 Example/package-lock.json
 package-lock.json
+
+# Built files
+tsconfig.tsbuildinfo
+index.js.flow
+index.d.ts
+index.js
