+++ Example/test/assetsTransformer.js
@@ -2,7 +2,7 @@
  * Prevent error from requiring asset during test.
  * https://github.com/facebook/jest/issues/2663
  */
-const path = require('path');
+import path from 'path';
 
 module.exports = {
   process(src, filename, config, options) {
