+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-audio-recorder-player",
-  "version": "2.3.0",
+  "version": "2.4.0",
   "description": "React Native Audio Recorder and Player.",
   "homepage": "https://github.com/dooboolab/react-native-audio-recorder-player",
   "main": "index.js",
@@ -8,7 +8,8 @@
   "postinstall": "dooboolab-welcome postinstall",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint -c .eslintrc.js '.'"
+    "build": "tsc && flowgen index.d.ts -o index.js.flow",
+    "lint": "eslint -c .eslintrc.js 'index.ts'"
   },
   "repository": {
     "type": "git",
@@ -32,7 +33,6 @@
     "dooboolab-welcome": "^1.1.1"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
     "@babel/cli": "^7.6.2",
     "@babel/core": "^7.6.2",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
@@ -40,6 +40,7 @@
     "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
     "@babel/plugin-transform-runtime": "7.6.2",
     "@babel/preset-typescript": "^7.6.0",
+    "@dooboo/eslint-config": "^0.1.3",
     "@babel/runtime": "^7.6.2",
     "@testing-library/react-native": "^4.0.9",
     "@types/jest": "^24.0.18",
@@ -50,18 +51,12 @@
     "@types/styled-components": "^4.1.19",
     "@typescript-eslint/eslint-plugin": "^2.2.0",
     "@typescript-eslint/parser": "^2.2.0",
+    "babel-eslint": "^10.0.1",
     "babel-jest": "24.9.0",
     "babel-preset-react-native": "^5.0.2",
     "codecov": "^3.5.0",
     "eslint": "^6.3.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.14.3",
-    "eslint-plugin-standard": "^4.0.1",
-    "eslint-plugin-typescript": "^0.14.0",
+    "flowgen": "^1.10.0",
     "prettier": "^1.18.2",
     "react-dom": "^16.10.0",
     "react-test-renderer": "^16.10.0",
