+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+  "compilerOptions": {
+    "lib": ["es2017"],
+    "jsx": "react-native",
+    "noResolve": false,
+    "noImplicitAny": false,
+    "removeComments": false,
+    "sourceMap": false,
+    "module": "es2015",
+    "moduleResolution": "node",
+    "target": "es5",
+    "skipLibCheck": true,
+    "allowSyntheticDefaultImports": true,
+    "declaration": true
+  },
+  "compileOnSave": true,
+  "exclude": ["node_modules"],
+  "include": ["./index.ts"]
+}
